添加可缩放的滚动条控件，横向纵向均使用该控件
可缩放的滚动条控件自身是认为总数据范围不变的，修改显示范围，记录显示范围变化率，如果需要改变总数据范围可以在外部使用时修改总数据范围和显示范围，再重绘；
使用放大倍数而不是最小像素作限制，避免差为0或负数或显示不下溢出的情况；
int行高缩放多次后相对比例差异越来越大，所以改成了float，横坐标暂时不需要改；

+++ TimeAxis.WinForms/Model/Row.cs
@@ -9,6 +9,9 @@ namespace TimeAxis
 {
     public abstract class Row
     {
-        public int Height { get; set; } = 25;
+        /// <summary>
+        /// 行高，因为涉及到整体缩放，为了保证多次缩放后相对比例不变所以用小数，比int精度更高
+        /// </summary>
+        public float Height { get; set; } = 25;
     }
 }
